Be verbose during Python package build
Debian requires the build log to show all gcc invocations with full
parameters. Also this makes it easier to debug problems related to build
flags. Right now for example, the build is failing when we enable
the PIE build flags and I need to see the command line to be able
to work on a fix.